requirements.txt
atproto~=0.0.55
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-Bcrypt~=1.0.1